simplify vindlu model code
