--- configs/multimodal/vindlu/vindlu_vqa_infer.py
@@ -1,91 +1,0 @@
-_base_ = ['../../_base_/default_runtime.py']
-
-video_root = 'data/msrvtt/msrvtt_2fps_224'
-anno_file_test = 'data/msrvtt/anno_downstream/msrvtt_qa_test.json'
-answer_list_file = 'data/msrvtt/anno_downstream/msrvtt_qa_answer_list.json'
-
-# model settings
-model = dict(
-    type='VindLUVQA',
-    data_preprocessor=dict(
-        type='ActionDataPreprocessor',
-        mean=[128],
-        std=[128],
-        format_shape='NCTHW'),
-    tokenizer=dict(
-        type='BertTokenizer',
-        pretrained_model_name_or_path='bert-base-uncased'),
-    vision_encoder=dict(
-        type='beit',
-        tem_config=dict(
-            num_frames=12,
-            temporal_model_block='timesformer',
-            temporal_model_position='last',
-            temporal_model_config=dict(input_dim=768),
-            use_temporal_position_embedding=True),
-        pretrained_model_name_or_path=
-        'microsoft/beit-base-patch16-224-pt22k-ft22k',
-        encoder_width=768,
-        add_ln=True),
-    text_encoder=dict(
-        type='XBertModel',
-        pretrained_model_name_or_path='bert-base-uncased',
-        encoder_width=768,
-        fusion_layer=9,
-        add_pooling_layer=False),
-    text_decoder=dict(
-        type='BertDecoder',
-        pretrained_model_name_or_path='bert-base-uncased',
-        encoder_width=768,
-        fusion_layer=0,
-        num_hidden_layers=3,
-        add_pooling_layer=True),
-    proj_dim=256,
-    temperature=0.07,
-    max_question_len=25,
-    max_answer_len=5,
-    num_ans_candidates=128,
-    gradient_checkpointing=True,
-    answer_list_path=answer_list_file)
-
-file_client_args = dict(io_backend='disk')
-
-test_pipeline = [
-    dict(type='DecordInit', **file_client_args),
-    dict(
-        type='SampleFrames',
-        clip_len=1,
-        frame_interval=1,
-        num_clips=12,
-        test_mode=True),
-    dict(type='DecordDecode'),
-    dict(type='Resize', scale=(224, 224), keep_ratio=False),
-    dict(type='FormatShape', input_format='NCHW'),
-    dict(
-        type='PackActionInputs',
-        algorithm_keys=(
-            'question',
-            'gt_answer',
-            'question_id',
-        ))
-]
-
-dataset_type = 'MSRVTT_VQA'
-
-test_dataloader = dict(
-    batch_size=32,
-    num_workers=8,
-    persistent_workers=True,
-    sampler=dict(type='DefaultSampler', shuffle=False),
-    dataset=dict(
-        type=dataset_type,
-        ann_file=anno_file_test,
-        pipeline=test_pipeline,
-        data_prefix=dict(video=video_root),
-    ))
-
-test_evaluator = dict(type='VQAAcc')
-test_cfg = dict(type='TestLoop')
-
-default_hooks = dict(
-    logger=dict(type='LoggerHook', interval=20, ignore_last=False), )
