+++ mmaction/models/multimodal/__init__.py
@@ -2,10 +2,11 @@
 from .beit3d import BeitModel3D
 from .tokenizer import BertTokenizer
 from .vindlu_ret import VindLURetrieval
-from .vindlu_ret_mc import VindLURetMC
+from .vindlu_ret_mc import VindLURetrievalMC
 from .vindlu_vqa import VindLUVQA
+from .xbert import BertDecoder, BertModel
 
 __all__ = [
-    'VindLUVQA', 'BertTokenizer', 'BeitModel3D', 'VindLURetMC',
-    'VindLURetrieval'
+    'VindLUVQA', 'BertTokenizer', 'BertModel', 'BertDecoder', 'BeitModel3D',
+    'VindLURetrievalMC', 'VindLURetrieval'
 ]
